ARM: tegra: cardhu - Enable HDMI
Enable the HDMI output and hotplug detection. Note that HDMI doesn't
work on Cardhu for some reason. I remember someone reporting that it
worked at some point in time, but that no longer seems to be true.

A potential issue might be that EMC clocks are too slow, due to them
being set to some safe default, which causes underflow for high HDMI
resolutions.

This can be made to work by modifying the display A and display B
latency allowance registers in the MC.

Signed-off-by: Thierry Reding <treding@nvidia.com>