drm/tegra: Statically assign pipe number
The driver previously relied on the probe order to assign a pipe number
to each head. With the introduction of panel support it is now possible
for the display controller driver to defer probing and therefore it can
happen that the heads are assigned the wrong pipe number, which causes
breakage later on because the wrong values end up being programmed into
registers of the various outputs.

Signed-off-by: Thierry Reding <treding@nvidia.com>