drm/tegra: Fix possible CRTC mask for RGB outputs
The mask of possible CRTCs that an output (DRM encoder) can be attached
to is relative to the position within the DRM device's list of CRTCs.
Deferred probing can cause this to not match the pipe number associated
with a CRTC. Use the newly introduced drm_helper_crtc_possible_mask() to
compute the mask by looking up the proper index of the given CRTC in the
list.

Signed-off-by: Thierry Reding <treding@nvidia.com>