PCI/AER: Make linux/aer.h standalone includable
The header file references u16, u32 and struct pci_dev types, but they
are not defined in the header nor does the header pull in the necessary
includes for them. This causes build breakage when the file is included
without any of the dependencies being satisfied from somewhere else.

Fix this by including linux/types.h (for u16 and u32) and by adding a
forward-declaration of struct pci_dev.

Signed-off-by: Thierry Reding <treding@nvidia.com>